One-file to rule them all - almost
