+++ Models/Post.php
@@ -1,24 +1,20 @@
 <?php
 
-namespace V\Models;
-
-interface IPost {
+interface IPost
+{
     public function link(): string;
 }
 
 /**
  * Class Post
- *
  * The Post Model
- *
- * @package V\Models
  */
-class Post implements IPost
+class Post extends Volcano implements IPost
 {
     public function __construct(
         protected string $filepath = '',
-    )
-    {}
+    ) {
+    }
 
     /**
      * Generate a post link.
@@ -28,7 +24,7 @@ class Post implements IPost
      */
     public function link(): string
     {
-        $noDirectoryName = explode('/', str_replace(POSTS, '', $this->filepath));
+        $noDirectoryName = explode('/', str_replace($this->getPath('posts'), '', $this->filepath));
         return str_replace('.md', '', str_replace('\\', '/', end($noDirectoryName)));
     }
 }
