--- index.php
@@ -1,5 +1,0 @@
-<?php
-
-define('ROOT', dirname(__FILE__));
-
-require 'core/config.php';
